pyproject.toml
[project]
name = "live-notebook-lm"
version = "0.1.0"
description = "LiveNotebookLM - NotebookLM with real-time voice interaction via Gemini Live API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "certifi>=2024.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
